+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/generictest/serializer/SerializerTestUUID.java
@@ -1,24 +1,23 @@
-package io.github.pylonmc.pylon.test.generictest;
+package io.github.pylonmc.pylon.test.generictest.serializer;
 
 import io.github.pylonmc.pylon.core.persistence.PylonPersistentDataContainer;
 import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers;
 import io.github.pylonmc.pylon.test.GenericTest;
 import io.github.pylonmc.pylon.test.TestAddon;
-import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.NamespacedKey;
 import org.bukkit.inventory.ItemStack;
 
+import java.util.UUID;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class SerializerTestLocation implements GenericTest {
+public class SerializerTestUUID implements GenericTest {
     @Override
     public void run() {
         NamespacedKey key = new NamespacedKey(TestAddon.instance(), "key");
-        var type = PylonSerializers.LOCATION;
-        Location value = new Location(TestAddon.testWorld, 5.0, 320.3, 38904.43443);
-        value.setPitch(3.0F);
-        value.setYaw(1.04F);
+        var type = PylonSerializers.UUID;
+        UUID value = UUID.randomUUID();
 
         ItemStack stack = new ItemStack(Material.ACACIA_BOAT);
         stack.editMeta(meta -> meta.getPersistentDataContainer()
