+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/generictest/serializer/SerializerTestVector.java
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.test.generictest;
+package io.github.pylonmc.pylon.test.generictest.serializer;
 
 import io.github.pylonmc.pylon.core.persistence.PylonPersistentDataContainer;
 import io.github.pylonmc.pylon.core.persistence.datatypes.PylonSerializers;
